CorretFilter2DLinebuffer
The line buffer logic was technically wrong. The line0, line1, line2 should be strictly aligned with the filtering kernel. Therefore, line0 should always be the row above the central row; line 1 is always the central row; line2 is always the row below the central row. The original code is super random.
The 'count' was not reset.
The previous error makes the output not consistent with the same input image.

Signed-off-by: Zhenyu Xu <zxu3@clemson.edu>